==> wgcna/1. clean-match-clean.ipynb <==
import pandas as pd
import numpy as np

microbiome=pd.read_table("../../haddad_data/raw/otu_table_haddad_fecal__expected_duration_6weeks__.txt", sep='\t', skiprows=1, index_col="#OTU ID")

metabolome=pd.read_table("../../haddad_data/raw/haddad_master_featuretable_unnorm.txt", sep='\t', header=0, index_col="#OTU ID")

metadata=pd.read_table("../../haddad_data/raw/haddad_master_metadata_fecal_w_alphametrices__expected_duration_6_weeks__.txt", sep='\t', header=0,index_col="#SampleID")

#clean micro
microbiome=microbiome.T
microbiome=microbiome.reset_index().rename(columns={"index":"#SampleID"})
microbiome=microbiome.set_index("#SampleID")

#clean metab
metabolome=metabolome.T
metabolome=metabolome.reset_index().rename(columns={"index":"#SampleID"})
metabolome=metabolome.set_index("#SampleID")

def match_samples(mic, met, md):
    mic_samp=set(mic.index)
    met_samp=set(met.index)
    md_samp=set(md.index)
    matched=sorted(list(mic_samp & met_samp & md_samp))
    mic=mic.ix[matched]
    met=met.ix[matched]
    md=md.ix[matched]
    return(mic, met, md)
    
    import unittest
    m1 = pd.DataFrame([[0, 0, 1],
                  [2, 2, 4],
                  [5, 5, 3]],
                  index=['s1', 's2', 's3'],
                  columns=['o1', 'o2', 'o3'])
    _mapping = pd.DataFrame([['a', 'control'],
                        ['b', 'control'],
                        ['c', 'diseased'],
                        ['d', 'diseased']],
                        index=['s1', 's2', 's3', 's4'],
                        columns=['Barcode', 'Treatment'])
    m2=pd.DataFrame([[2, 4, 4],
                 [5, 3, 3],
                 [0, 0, 1]],
                index=['s2', 's3', 's4'],
                columns=['c1', 'c2', 'c3'])
    assert(match_samples(m1, m2, _mapping)[0].index==['s2','s3'])

microbiome, metabolome, metadata=match_samples(microbiome, metabolome, metadata)

print(microbiome.shape)
print(metabolome.shape)
print(metadata.shape)

#### filter zero features


frem=[]
for otu in microbiome.columns:
    ll=np.array(microbiome[otu])
    if np.count_nonzero(ll) == 0:
        frem.append(otu)
microbiome=microbiome.drop(frem, axis=1)

frem=[]
for mol in metabolome.columns:
    ll=np.array(metabolome[mol])
    if np.count_nonzero(ll) == 0:
        frem.append(mol)
metabolome=metabolome.drop(frem, axis=1)

print(microbiome.shape)
print(metabolome.shape)
print(metadata.shape)

microbiome.head(3)

metabolome.head(3)

metadata.head(3)

microbiome.T.to_csv("../../haddad_data/otu_6week_matched_scolumns.txt", sep='\t')
microbiome.to_csv("../../haddad_data/otu_6week_matched_srows.txt", sep='\t')

metabolome.T.to_csv("../../haddad_data/metab_6weeks_matched_scolumns.txt", sep='\t')
metabolome.to_csv("../../haddad_data/metab_6weeks_matched_srows.txt", sep='\t')

metadata.to_csv("../../haddad_data/metadata_matched.txt", sep='\t')

#### for metaboanalyst

These dataframes are matched in samples and order of samples. So we can simply add metadata column for our classifier like this:

microbiome["category"]=metadata["exposure_type"]
metabolome["category"]=metadata["exposure_type"]

microbiome.to_csv("../../haddad_data/otu_6week_matched_forclassifier.csv", sep=',')
metabolome.to_csv("../../haddad_data/metab_6weeks_matched_forclassifier.csv", sep=',')
