==> wgcna/1a.features-to-ids.ipynb <==
import pandas as pd
import random
random.seed(2017)

metabft=pd.read_table("../../classifier_filter/metab_rf_filter_srows.txt", sep='\t', dtype=str, header=0)#index_col="#feature ID")
microft=pd.read_table("../../classifier_filter/micro_rf_filter_srows.txt", sep='\t', dtype=str, header=0)
gnps=pd.read_table("./gnps_file.tsv", sep='\t', dtype=str, header=0)
gg97=pd.read_table("/Users/priya/Desktop/kd-research/gg_13_8_otus/taxonomy/97_otu_taxonomy.txt",sep='\t', dtype=str, header=0, index_col="#OTU ID")

metabft.head(3)

microft.head(3)

gnps.head(3)

gg97.head(3)

## fuzzy conversion of MS1 features to MS2-based identification

mztol=15
rttol=20
for ft in metabft.columns[1:]:
    
    feature_mzdiff=[]
    mz=ft.split('_')[0]
    rt=ft.split('_')[1]
    for idx in gnps.index:
        mzdiff= abs(float(mz)-float(gnps.ix[idx, 'Precursor_MZ'])) 
        rtdiff= abs(float(mz)-float(gnps.ix[idx, 'RT_Query']))
        if mzdiff <=mztol and rtdiff <=rttol:
            feature_mzdiff.append((gnps.ix[idx, 'Compound_Name'], mzdiff))
    if feature_mzdiff==[]:
        continue
    else:
        sfeature_mzdiff=sorted(feature_mzdiff, key = lambda element : element[1])
        metabft=metabft.rename(columns={ft:sfeature_mzdiff[0][0]+"_"+str(mz)[0:5]+"_"+str(rt)[0:5]})

metabft.head(3)

metabft.to_csv("./metabft_w_ids_srows.txt", sep='\t', index=False)

## converting OTU IDs to taxonomy

otu_tax={}
for otu in gg97.index:
    otu_tax[otu]=gg97.ix[otu,"taxonomy"]

for otus in microft.columns[1:]:
    microft=microft.rename(columns={otus:otu_tax[int(otus)]+"id"+otus})

microft.head()

microft.to_csv("./microft_w_ids_srows.txt", sep='\t', index=False)
