==> wgcna/2. keep-imp-features.ipynb <==
import pandas as pd
import scipy.stats as ss
import numpy as np
% matplotlib inline

microbiome=pd.read_table("../otu_6week_matched_srows.txt", sep='\t', index_col="#SampleID", dtype=str)
metabolome=pd.read_table("../metab_6weeks_matched_srows.txt", sep='\t', index_col="#SampleID", dtype=str)

print(microbiome.shape)
print(metabolome.shape)

rf_microb=pd.read_table("../process1/rf_otu_proto.csv", sep=',', index_col=0, dtype=str)
rf_metab=pd.read_table("../process1/rf_metab_proto.csv", sep=',', index_col=0, dtype=str)

thresh=0
rf_microb["MeanDecreaseAccuracy"]=pd.to_numeric(rf_microb["MeanDecreaseAccuracy"])
rf_metab["MeanDecreaseAccuracy"]=pd.to_numeric(rf_metab["MeanDecreaseAccuracy"])

impbac=[bac for bac in rf_microb.index if rf_microb.ix[bac,"MeanDecreaseAccuracy"]>thresh]
impmol=[mol for mol in rf_metab.index if rf_metab.ix[mol,"MeanDecreaseAccuracy"]>thresh]

print(len(impbac))
print(len(impmol))

mic=pd.DataFrame(index=microbiome.index)
for bac in impbac:
    mic[str(bac)]=microbiome[str(bac)]

met=pd.DataFrame(index=metabolome.index)
for mol in impmol:
    met[str(mol)]=metabolome[str(mol)]

mic.head(3)

met.head(3)

print(mic.shape)
print(met.shape)

## normalize feature in met
for feature in met.columns:
    met[feature]=pd.to_numeric(met[feature])
    met[feature]=met[feature]/float(sum(met[feature]))

##normalize sample in mic and met
for samp in mic.index:
    mic.ix[samp]=pd.to_numeric(mic.ix[samp])
    met.ix[samp]=pd.to_numeric(met.ix[samp])
    mic.ix[samp]=mic.ix[samp]/float(sum(mic.ix[samp]))
    met.ix[samp]=met.ix[samp]/float(sum(met.ix[samp]))

##test
ss.spearmanr(np.array(met['222.066842105263_385.199902286792']),np.array(mic['268733']))

# #pairwise correlations taking too long
# pairs=[]
# for mol in met.columns:
#     for bac in mic.columns:
#         res=ss.spearmanr(np.array(met[mol]),np.array(mic[bac]))
#         pairs.append(res)

met.to_csv("../metab_rf_filter_srows.txt", sep='\t')
met.T.to_csv("../metab_rf_filter_scolumns.txt", sep='\t')
mic.to_csv("../micro_rf_filter_srows.txt", sep='\t')
mic.T.to_csv("../micro_rf_filter_scolumns.txt", sep='\t')

## pairwise correlations between
