==> wgcna/2.Haddad_WGCNA_modules.ipynb <==
options(warn = -1)
library(WGCNA)
options(stringsAsFactors=FALSE)
enableWGCNAThreads()

load(file='metab_mapping.RData')
load(file='microb_mapping.RData')

# choose soft-thresholding power for metab
powers_metab = c(c(1:10), seq(12, 20, by=2)) 
sft_metab = pickSoftThreshold(dat_metab, powerVector=powers_metab, verbose=5) 

# cex1 = 0.9 # size of red number 
# plot(sft_metab$fitIndices[,1], -sign(sft_metab$fitIndices[,3])*sft_metab$fitIndices[,2],
#      xlab='Metab Soft Threshold (power)', ylab = 'Scale Free Topology Model Fit',
#     type='n', main=paste('Scale independence'))
# text(sft_metab$fitIndices[,1], -sign(sft_metab$fitIndices[,3])*sft_metab$fitIndices[,2],
#                               labels=powers_metab, cex=cex1, col='red')
# abline(h=0.4, col='red') # choose power=4, the lowest power for which the curve flattens out

# choose soft-thresholding power for Microb
powers_microb = c(c(1:10), seq(12, 20, by=2))
sft_microb = pickSoftThreshold(dat_microb, powerVector=powers_microb, verbose=5) 

# cex1 = 0.9 # size of red number 
# plot(sft_microb$fitIndices[,1], -sign(sft_microb$fitIndices[,3])*sft_microb$fitIndices[,2],
#      xlab='microb Soft Threshold (power)', ylab = 'Scale Free Topology Model Fit',
#     type='n', main=paste('Scale independence'))
# text(sft_microb$fitIndices[,1], -sign(sft_microb$fitIndices[,3])*sft_microb$fitIndices[,2],
#                               labels=powers_microb, cex=cex1, col='red')
# abline(h=0.4, col='red') # choose power=4, the lowest power for which the curve flattens out

net_metab = blockwiseModules(dat_metab, power=4, TOMtype='unsigned', minModuleSize=10, reassignThreshold=0, 
                                mergeCutHeight=0.25, # numericLabels=TRUE,
                                pamRespectsDendro=FALSE, saveTOMS=TRUE, saveTOMFileBase='HadadMouseTOM')

net_microb = blockwiseModules(dat_microb, power=4, TOMtype='unsigned', minModuleSize=30, reassignThreshold=0, 
                                mergeCutHeight=0.25, # numericLabels=TRUE,
                                pamRespectsDendro=FALSE, saveTOMS=TRUE, saveTOMFileBase='HadadMouseTOM')

print(table(net_metab$colors))
print(table(net_microb$colors))

mergedColors_metab = labels2colors(net_metab$colors)
pdf('dendrogram_metab.pdf', width=50, height=30)
plotDendroAndColors(net_metab$dendrograms[[1]], mergedColors_metab[net_metab$blockGenes[[1]]], # net$dendrograms[[1]] is the distance matrix
                   'Module Colors', dendroLabels=FALSE, hang=0.03,
                   addGuide=TRUE, guideHang=0.05, main='Haddad Metab dendrogram and module')
dev.off()

mergedColors_microb = labels2colors(net_microb$colors)
pdf('dendrogram_microb.pdf', width=50, height=30)
plotDendroAndColors(net_microb$dendrograms[[1]], mergedColors_microb[net_microb$blockGenes[[1]]], # net$dendrograms[[1]] is the distance matrix
                   'Module Colors', dendroLabels=FALSE, hang=0.03,
                   addGuide=TRUE, guideHang=0.05, main='Haddad microb dendrogram and module')
dev.off()

# save result from clustering
moduleLabels_metab = net_metab$colors
moduleColors_metab = labels2colors(net_metab$colors)
MEs_metab = net_metab$MEs
tree_metab = net_metab$dendrograms[[1]]

moduleLabels_microb = net_microb$colors
moduleColors_microb = labels2colors(net_microb$colors)
MEs_microb = net_microb$MEs
tree_microb = net_microb$dendrograms[[1]]

save(moduleLabels_metab, moduleColors_metab, MEs_metab, tree_metab, 
     file='Haddad_metab_modules.RData')
save(moduleLabels_microb, moduleColors_microb, MEs_microb, tree_microb, 
     file='Haddad_microb_modules.RData')
