==> wgcna/4.features-to-ids.ipynb <==
import pandas as pd

ft=pd.read_table("./sotu_metab_corrmat_tq_br.tsv", sep='\t', dtype=str, header=0) #index_col="#feature ID")
gnps=pd.read_table("./gnps_file.tsv", sep='\t', dtype=str, header=0)
gg97=pd.read_table("/Users/priya/Desktop/kd-research/gg_13_8_otus/taxonomy/97_otu_taxonomy.txt",sep='\t', dtype=str, header=0, index_col="#OTU ID")

ft.head(3)

gnps.head(3)

gg97.head(3)

## fuzzy conversion of MS1 features to MS2-based identification

mztol=15
rttol=20
for idx1 in ft.index:
    
    feature_mzdiff=[]
    mz=ft.ix[idx1, "#feature ID"].split('_')[0]
    rt=ft.ix[idx1, "#feature ID"].split('_')[1]
    for idx2 in gnps.index:
        mzdiff= abs(float(mz)-float(gnps.ix[idx2, 'Precursor_MZ'])) 
        rtdiff= abs(float(mz)-float(gnps.ix[idx2, 'RT_Query']))
        if mzdiff <=mztol and rtdiff <=rttol:
            feature_mzdiff.append((gnps.ix[idx2, 'Compound_Name'], mzdiff))
    if feature_mzdiff==[]:
        continue
    else:
        sfeature_mzdiff=sorted(feature_mzdiff, key = lambda element : element[1])
        ft.ix[idx1, "#feature ID"]=sfeature_mzdiff[0][0]

ft.head(5)

## converting OTU IDs to taxonomy

otu_tax={}
for otu in gg97.index:
    otu_tax[otu]=gg97.ix[otu,"taxonomy"]

for otus in ft.columns[1:]:
    ft=ft.rename(columns={otus:otu_tax[int(otus)]+"id"+otus})

ft.head()

ft.to_csv("mm_features_pairwise_corr_tq_br.tsv", sep='\t')

### test: looking for adlercreutzia in filtered otu table

import pandas as pd
feature_microb=pd.read_table("../../classifier_filter/micro_rf_filter_srows.txt", sep='\t', dtype=str)

adler=[]
for otu in gg97.index:
    if "g__Adlercreutzia" in gg97.ix[otu, "taxonomy"]:
        adler.append(str(otu))

adler

for feature in feature_microb.columns:
    if feature in adler:
        print(feature)

### test: looking for enterolactone and enterodiol

ft=pd.read_table("../../classifier_filter/metab_rf_filter_scolumns.txt", sep='\t', dtype=str)

ft.head(3)

#make list of interesting mzrts from gnps hit file
mzrts=[]
for idx in gnps.index:
    if gnps.ix[idx,"Compound_Name"] == "Enterolactone":
        mzrts.append((gnps.ix[idx,'Precursor_MZ'], gnps.ix[idx,'RT_Query']))

for idx in gnps.index:
    if gnps.ix[idx,"Compound_Name"] == "Enterodiol":
        mzrts.append((gnps.ix[idx,'Precursor_MZ'], gnps.ix[idx,'RT_Query']))

mzrts

feature_id=[]
mztol=15
rttol=20
for idx1 in ft.index:
    mz=ft.ix[idx1, "#OTU ID"].split('_')[0]
    rt=ft.ix[idx1, "#OTU ID"].split('_')[1]
#     print mz
    for elem in mzrts:
#         print(elem[0])
        mzdiff= abs(float(mz)-float(elem[0])) 
        rtdiff= abs(float(rt)-float(elem[1]))
        if mzdiff <=mztol and rtdiff <=rttol:
            print(ft.ix[idx1, "#OTU ID"])
